chore: Update cache cleanup workflow (#37)
* chore: Update cache cleanup workflow

Update the cache cleanup workflow to run on ubuntu-latest and rename it to 'Cleanup Cache'.
Also, the workflow is triggered on pull request close and workflow dispatch.
Finally, it uses the openMF/mifos-mobile-github-actions cache cleanup action.

* chore: Update cache cleanup workflow

Update the cache cleanup workflow to run on ubuntu-latest and rename it to 'Cleanup Cache'.
Also, the workflow is triggered on pull request close and workflow dispatch.
Finally, it uses the openMF/mifos-mobile-github-actions cache cleanup action.